requirements.txt
kafka-python==2.0.2
fake_web_events== 0.2.5